perf - log startup timings in web same as desktop
