+++ src/vs/workbench/contrib/performance/electron-sandbox/performance.contribution.ts
@@ -7,12 +7,13 @@ import { LifecyclePhase } from 'vs/workbench/services/lifecycle/common/lifecycle
 import { Registry } from 'vs/platform/registry/common/platform';
 import { Extensions, IWorkbenchContributionsRegistry } from 'vs/workbench/common/contributions';
 import { StartupProfiler } from './startupProfiler';
-import { StartupTimings } from './startupTimings';
+import { NativeStartupTimings } from './startupTimings';
 import { RendererProfiling } from 'vs/workbench/contrib/performance/electron-sandbox/rendererAutoProfiler';
 import { IConfigurationRegistry, Extensions as ConfigExt } from 'vs/platform/configuration/common/configurationRegistry';
 import { localize } from 'vs/nls';
 import { applicationConfigurationNodeBase } from 'vs/workbench/common/configuration';
 
+// -- auto profiler
 
 Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench).registerWorkbenchContribution(
 	RendererProfiling,
@@ -29,7 +30,7 @@ Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench).registerWorkb
 // -- startup timings
 
 Registry.as<IWorkbenchContributionsRegistry>(Extensions.Workbench).registerWorkbenchContribution(
-	StartupTimings,
+	NativeStartupTimings,
 	LifecyclePhase.Eventually
 );
